backend/requirements.txt
# Python dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# ML and Computer Vision
torch==2.5.1
torchvision==0.20.1
git+https://github.com/openai/CLIP.git
opencv-python==4.8.1.78
pillow==10.1.0
mediapipe==0.10.21
scikit-learn==1.3.2

# Data Processing
pandas==2.1.3
numpy==1.26.2

# CLIP dependencies
ftfy==6.1.3
regex==2023.10.3
tqdm==4.66.1
